+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/camera/openCV/AprilTags/AprilTag.java
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode.openCV.AprilTags;
+package org.firstinspires.ftc.teamcode.camera.openCV.AprilTags;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;