+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/camera/openCV/AprilTags/AprilTagAutoInitDetectionExample.java
@@ -1,5 +1,6 @@
-package org.firstinspires.ftc.teamcode.openCV.AprilTags;
+package org.firstinspires.ftc.teamcode.camera.openCV.AprilTags;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -12,7 +13,7 @@ import org.openftc.easyopencv.OpenCvCameraRotation;
 import java.util.ArrayList;
 
 @TeleOp
-//@Disabled
+@Disabled
 public class AprilTagAutoInitDetectionExample extends LinearOpMode {
     OpenCvCamera camera;
     AprilTagDetectionPipeline aprilTagDetectionPipeline;
@@ -145,6 +146,7 @@ public class AprilTagAutoInitDetectionExample extends LinearOpMode {
         while (opModeIsActive()) {
             sleep(20);
         }
+
     }
 
     void tagToTelemetry(AprilTagDetection detection) {