+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/camera/openCV/AprilTags/AprilTagDetectionPipeline.java
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode.openCV.AprilTags;
+package org.firstinspires.ftc.teamcode.camera.openCV.AprilTags;
 
 import org.opencv.calib3d.Calib3d;
 import org.opencv.core.CvType;