+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/camera/openCV/InternalCamera2AdvancedFeaturesExample.java
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode.openCV;
+package org.firstinspires.ftc.teamcode.camera.openCV;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;