+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/changeLog.md
@@ -1,3 +1,8 @@
+1.1.0 June 12th,2023
+Added config.md and a template
+OpenCV tests working (and april tags) and will switch with vuforia as well
+blink class completely tested and working
+
 1.0.1 June 5th,2023
 Finished Potentiometer and Limit Switches
 New blink class with everything to control the Blinkin driver